Extracted linkUnlinkFilter and added a test for it.
