+++ index.js
@@ -4,6 +4,7 @@ var TaskPlanner   = require('taskplanner')
   , assert        = require('assert')
   , xtend         = require('xtend')
   , allParentsIds = require('./lib/allParentsIds')
+  , linkFilter    = require('./lib/linkUnlinkFilter')
   , defaults      = {
         mode: 'quick'
       , noLinkUnlinkRemove: false
@@ -16,8 +17,6 @@ function planner(origin, dest, opts)  {
     , cmds      = generateCommands(origin, dest)
 
     , state     = _.cloneDeep(origin)
-
-    , skipNext  = false
 
     , result
 
@@ -58,29 +57,7 @@ function planner(origin, dest, opts)  {
   })
 
   if (!opts.noLinkUnlinkRemove)
-    result = result.filter(function(cmd, i, cmds) {
-
-      if (skipNext) {
-        skipNext = false
-        return false
-      }
-
-      if (!cmds[i + 1])
-        return true
-
-      if (cmds[i + 1].id !== cmd.id)
-        return true
-
-      var unlinkLink = (cmd.cmd === 'unlink' && cmds[i + 1].cmd === 'link')
-        , linkUnlink = (cmd.cmd === 'link' && cmds[i + 1].cmd === 'unlink')
-
-      if (linkUnlink || unlinkLink) {
-        skipNext = true
-        return false
-      }
-
-      return true
-    })
+    result = linkFilter(result)
 
   return result
 }
